Refactor write and decode
This refactors the code involved with writing and decoding U-Boot
environments. The original purpose was to reduce the amount of
binary garbage that got generated on load, but this contains some
cleanup as well.

On one system, then reduces the garbage created by about 2 MB. This
garbage could be cleaned up manually by calling
`:erlang.garbage_collect/1`, but it stayed around for a long time. Now
the total memory for the processes that had the garbage is in the 40 KB
range without a garbage collect.

The previous commit that uses iodata likely had an effect as well.